refactor(bridge): extract presence events
